QtBridge-Rust: Fix cargo clippy warnings and errors
Resolved warnings and errors from cargo clippy in
the qt_gen_common_src directory.

Task-number: QTBUG-143445

+++ qtbridge/qt_gen/impl/qt_gen_common_src/case_conv.rs
@@ -31,7 +31,7 @@ pub fn camel_to_snake(input: &str) -> String {
             if pos - last_capital_pos > 1 && last_ch != '_' {
                 result.push('_');
             }
-            last_capital_pos = pos as i32;
+            last_capital_pos = pos;
         }
         result.push(ch_lc);
         last_ch = ch_lc;
